+++ nas/__init__.py
@@ -1,0 +1,11 @@
+###################################################################################################
+#
+# Copyright (C) 2020 Maxim Integrated Products, Inc. All Rights Reserved.
+#
+# Maxim Integrated Products, Inc. Default Copyright Notice:
+# https://www.maximintegrated.com/en/aboutus/legal/copyrights.html
+#
+###################################################################################################
+"""
+NAS package
+"""
